+++ spec/app/models/blog.rb
@@ -1,0 +1,5 @@
+class Blog
+  include Mongoid::Document
+  has_many :posts, :validate => false
+  default_scope includes(:posts)
+end