+++ CHANGELOG.md
@@ -10,6 +10,9 @@ For instructions on upgrading to newer versions, visit
 * \#1994 `dependent: :delete` only hits the database once now for one to
   many and many to many relations instead of once for each document.
 
+* \#1987 Don't double-insert documents into identity map when eager loading
+  twice inside the unit of work.
+
 * \#1976 Don't execute eager load queries when base query is empty.
 
 * \#1953 Uniqueness validation now works on localized fields.